fix!: Remove path as a Parameter Supplied by generateNotification
Triggering of a notification is specific to the path which is supplied at the time of triggering, generateNotification does not have to provide the same path back as a parameter. With the path already supplied to the trigger function, the negotiated fields (which specific to the path) are sufficient to tailor the notification.