feat!: Simplified RFC822 templates
A simplified RFC822 template that requires only the specific properties to be rendered are passed to it, rather than the entire response object.

A new `defaultNotification` function is exposed by the PREP middleware that consumer may use instead of the template, which also now serves as the default value for `generateNotification` argument in the `trigger` function.