fix: Remove Delimiter Trailing LineBreak (#3)
We do not send a trailing internet linebreak immediately with the notification boundary delimiter. Instead each new notification starts with a linebreak.

This allows the server to close a stream by immediately sending trailing dashes without an empty message. Clients benefit from not having to buffer bytes for the next notification.

Signed-off-by: Rahul Gupta <CxRes@users.noreply.github.com>